mgmt: mcumgr: transport: Move BM UART code out into transport file
Allows including the BM UART transport in other applications and
updates the UART MCUmgr sample to use this transport library
