+++ samples/mcumgr/uart_mcumgr/CMakeLists.txt
@@ -10,13 +10,5 @@ find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(uart_mcumgr)
 
 zephyr_include_directories(include)
-zephyr_include_directories(${ZEPHYR_BASE}/subsys/mgmt/mcumgr/util/include)
-zephyr_include_directories(${ZEPHYR_BASE}/subsys/mgmt/mcumgr/grp/img_mgmt/include)
 zephyr_include_directories(${ZEPHYR_BASE}/subsys/mgmt/mcumgr/transport/include)
-target_sources(app PRIVATE
-  src/uart_mcumgr.c
-  src/smp_uart.c
-  ${ZEPHYR_BASE}/subsys/mgmt/mcumgr/transport/src/serial_util.c
-)
-zephyr_linker_sources(SECTIONS mcumgr_handler.ld)
-zephyr_link_libraries(MCUBOOT_BOOTUTIL)
+target_sources(app PRIVATE src/main.c)
