+++ samples/mcumgr/uart_mcumgr/Kconfig
@@ -6,39 +6,7 @@
 
 menu "UART MCUmgr sample"
 
-config UART_MCUMGR_RX_BUF_SIZE
-	int "Size of receive buffer for mcumgr fragments received over UART, in bytes"
-	default 128
-	help
-	  Specifies the size of the mcumgr UART receive buffer, in bytes. This
-	  value must be large enough to accommodate any line sent by an mcumgr
-	  client.
-
-config UART_MCUMGR_RX_BUF_COUNT
-	int "Number of receive buffers for mcumgr fragments received over UART"
-	default 8
-	help
-	  Specifies the number of the mcumgr UART receive buffers.  Receive
-	  buffers hold received mcumgr fragments prior to reassembly.  This
-	  setting's value must satisfy the following relation:
-	  UART_MCUMGR_RX_BUF_COUNT * UART_MCUMGR_RX_BUF_SIZE >=
-	  MCUMGR_TRANSPORT_UART_MTU
-
-config UARTE_HWFC
-	bool "UARTE HWFC"
-	help
-	  Enable hardware flow control on the UARTE peripheral used by the application.
-
-config UARTE_PARITY
-	bool "UARTE parity"
-	help
-	  Enable parity on the UARTE peripheral used by the application.
-
-config UARTE_IRQ_PRIO
-	int "UARTE IRQ priority"
-	default 3
-
-module=UART_MCUMGR_SAMPLE
+module=APP_UART_MCUMGR_SAMPLE
 module-str=UART MCUmgr Sample
 source "$(ZEPHYR_BASE)/subsys/logging/Kconfig.template.log_config"
