+++ samples/mcumgr/uart_mcumgr/prj.conf
@@ -20,6 +20,7 @@ CONFIG_CRC=y
 CONFIG_REBOOT=y
 CONFIG_NCS_BM_MCUMGR=y
 CONFIG_MCUMGR_TRANSPORT_REASSEMBLY=y
+CONFIG_MCUMGR_TRANSPORT_BM_UART=y
 CONFIG_MCUMGR_MGMT_NOTIFICATION_HOOKS=y
 CONFIG_MCUMGR_GRP_IMG=y
 CONFIG_MCUMGR_GRP_IMG_SLOT_INFO=y
