+++ scripts/ci/license_allow_list.yaml
@@ -54,8 +54,8 @@ Apache-2.0: |
   ^nrf-bm/subsys/mgmt/mcumgr/transport/src/smp.c
   ^nrf-bm/subsys/mgmt/mcumgr/util/CMakeLists.txt
   ^nrf-bm/sysbuild/Kconfig.bm
-  ^nrf-bm/samples/mcumgr/uart_mcumgr/src/smp_uart.c
-  ^nrf-bm/samples/mcumgr/uart_mcumgr/src/uart_mcumgr.c
+  ^nrf-bm/subsys/mgmt/mcumgr/transport/src/bm_uart_mcumgr.c
+  ^nrf-bm/subsys/mgmt/mcumgr/transport/src/smp_uart.c
   ^nrf-bm/lib/bm_zms/
   ^nrf-bm/include/bm_zms.h
   ^nrf-bm/tests/lib/bm_zms/
