+++ subsys/bluetooth/services/ble_mcumgr/CMakeLists.txt
@@ -8,5 +8,5 @@ zephyr_sources(mcumgr.c)
 zephyr_include_directories(${ZEPHYR_BASE}/subsys/mgmt/mcumgr/util/include)
 zephyr_include_directories(${ZEPHYR_BASE}/subsys/mgmt/mcumgr/grp/img_mgmt/include)
 zephyr_include_directories(${ZEPHYR_BASE}/subsys/mgmt/mcumgr/transport/include)
-zephyr_linker_sources(SECTIONS mcumgr_handler.ld)
+zephyr_linker_sources(SECTIONS ${ZEPHYR_NRF_BM_MODULE_DIR}/subsys/mgmt/mcumgr/transport/mcumgr_handler.ld)
 zephyr_link_libraries(MCUBOOT_BOOTUTIL)
