--- subsys/bluetooth/services/ble_mcumgr/mcumgr_handler.ld
@@ -1,1 +1,0 @@
-ITERABLE_SECTION_ROM(mcumgr_handler, Z_LINK_ITERABLE_SUBALIGN)
