+++ subsys/mgmt/mcumgr/transport/CMakeLists.txt
@@ -1,6 +1,6 @@
 #
 # Copyright (c) 2018-2021 mcumgr authors
-# Copyright (c) 2022 Nordic Semiconductor ASA
+# Copyright (c) 2022-2025 Nordic Semiconductor ASA
 #
 # SPDX-License-Identifier: Apache-2.0
 #
@@ -14,8 +14,12 @@ zephyr_library_sources_ifdef(CONFIG_MCUMGR_TRANSPORT_REASSEMBLY
   ${ZEPHYR_BASE}/subsys/mgmt/mcumgr/transport/src/smp_reassembly.c
 )
 
-#if(CONFIG_MCUMGR_TRANSPORT_SHELL OR CONFIG_MCUMGR_TRANSPORT_UART)
-#  zephyr_library_sources(src/serial_util.c)
-#endif()
+zephyr_library_sources_ifdef(CONFIG_MCUMGR_TRANSPORT_BM_UART
+  src/smp_uart.c
+  src/bm_uart_mcumgr.c
+  ${ZEPHYR_BASE}/subsys/mgmt/mcumgr/transport/src/serial_util.c
+)
 
-zephyr_include_directories(${ZEPHYR_BASE}/subsys/mgmt/mcumgr/transport/include)
+zephyr_include_directories(${ZEPHYR_BASE}/subsys/mgmt/mcumgr/transport/include include)
+zephyr_library_include_directories(${ZEPHYR_BASE}/subsys/mgmt/mcumgr/util/include)
+zephyr_linker_sources(SECTIONS mcumgr_handler.ld)
