+++ subsys/mgmt/mcumgr/transport/Kconfig
@@ -37,3 +37,5 @@ config MCUMGR_TRANSPORT_NETBUF_SIZE
 module = MCUMGR_TRANSPORT
 module-str = mcumgr_transport
 source "subsys/logging/Kconfig.template.log_config"
+
+rsource "Kconfig.bm_uart"
