+++ subsys/mgmt/mcumgr/transport/Kconfig.bm_uart
@@ -1,0 +1,57 @@
+# Copyright Nordic Semiconductor ASA 2025. All rights reserved.
+# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
+
+# The Kconfig file is dedicated to the BM UART transport of MCUmgr
+# subsystem and provides Kconfig options to control aspects of
+# the transport.
+#
+# Options defined in this file should be prefixed:
+#  MCUMGR_TRANSPORT_BM_UART_
+
+menuconfig MCUMGR_TRANSPORT_BM_UART
+	bool "Bare Metal UART MCUmgr SMP transport"
+	depends on BASE64
+	depends on CRC
+	help
+	  Enables handling of SMP commands received over UART.
+
+if MCUMGR_TRANSPORT_BM_UART
+
+config MCUMGR_TRANSPORT_BM_UART_UARTE_HWFC
+	bool "UARTE HWFC"
+	help
+	  Enable hardware flow control on the UARTE peripheral used by the application.
+
+config MCUMGR_TRANSPORT_BM_UART_UARTE_PARITY
+	bool "UARTE parity"
+	help
+	  Enable parity on the UARTE peripheral used by the application.
+
+config MCUMGR_TRANSPORT_BM_UART_UARTE_IRQ_PRIO
+	int "UARTE IRQ priority"
+	default 3
+
+config MCUMGR_TRANSPORT_BM_UART_RX_BUF_SIZE
+	int "Size of receive buffer for mcumgr fragments received over UART, in bytes"
+	default 128
+	help
+	  Specifies the size of the mcumgr UART receive buffer, in bytes. This
+	  value must be large enough to accommodate any line sent by an mcumgr
+	  client.
+
+# Duplicate Zephyr symbol without console dependency to allow usage of Zephyr header file
+config UART_MCUMGR_RX_BUF_SIZE
+	int
+	default MCUMGR_TRANSPORT_BM_UART_RX_BUF_SIZE
+
+config MCUMGR_TRANSPORT_BM_UART_RX_BUF_COUNT
+	int "Number of receive buffers for mcumgr fragments received over UART"
+	default 8
+	help
+	  Specifies the number of the mcumgr UART receive buffers.  Receive
+	  buffers hold received mcumgr fragments prior to reassembly.  This
+	  setting's value must satisfy the following relation:
+	  UART_MCUMGR_RX_BUF_COUNT * UART_MCUMGR_RX_BUF_SIZE >=
+	  MCUMGR_TRANSPORT_UART_MTU
+
+endif # MCUMGR_TRANSPORT_BM_UART
