+++ subsys/mgmt/mcumgr/transport/mcumgr_handler.ld
